x y z all have basic functionality. dual z driver different step rates sorted.
was not letting marlin be as clever as it is already designed to.

fix was configuring second z motor driver as Z2 instead of E1.

whine eliminated - was related to erroneously addressing Z2 as E1. Whine caused by improper initialization of the TMC drivers, keeping them in spreadcycle instead of stealthchop.